major code refactoring
refactored the code into logical subsystems.
There are now 2 types of soucefiles.
The first are generic device drivers, like can, spi or usart.
The second type of sourcefiles contains logic for one special
task (like encoder / current_measurement).

The firmware is now divided into five seperate parts:
- encoder reading
- current measurement
- protocoll parsing / state updating
- controllers (speed/position-pid...)
- H-Bridge controll, eg PWM generation

Note this is an unsable development branch, the firmware compiles,
and the basic test seem to work, but as this was an very heavy
refactoring, there is no garantee that everything is still working.