rewrite of i2c driver
rewrote the i2c driver on an register basis and
ignored the whole "state" thing from the documentation.
The new i2c driver seems to be more robust than the
old one. It was tested, by letting it run for an hour.
It did not hang in this time.

Note, this version still includes debug