added class to represent an encoder
The new class can be used, to feed raw encoder values
continuously to it. The class itself will generate an
absolute position from it.
An other feature is, that the class can deal with "lost"
encoder ticks. As a drawback to this, it must know the
wrap value of the encoder.
It is also possible, to set an zero position.