major rewrite of the i2c interface code
The i2c interface had, because of it's asynchronous nature, problems
if two different drivers tried to access it simultaneously.
Therefore the interface was redesigned to code with this issue.
One can now get a handle to an i2c-Bus which will be used for the
read and write operations. Internally the writes and reads will
be queued and be properly returned to the owner of the handle.

This change is needed to support two different temperature sensors on
the same i2c bus.