more bugfixes for i2c driver
There was a problem that the command processing was not triggered if one
was sending new command, leading to the issue that the bus could lock
up if all users were just sending.

Additionally there was a problem that the hasResult flag was not properly
cleared, leading to an issue were one bus user could issue multiple
commands in a row.