Total refactoring of the firmware code
The whole library was re-factored, to be 'platform independent'
This means that the state machine, the protocol, the encoder interface
and the controllers can now be used as a standalone library.

For every platform that should be supported one needs to implement
a set of functions.

For now this code is alpha and only the stub implementation es even
compiling. Nothing has been tested yet.