Firmware: Fix deadlock in CAN_SendMessage
There was a deadlock that stroke in the case that there was
no mailbox and the usart send buffer was full. Also usart
needed to be in interrupt mode for this to happen.
The deadlock was caused by the printf function waiting for
the send to finish. But as interrupts were disabled the
usart send buffer never got emptied.